+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2020-02-16
+### Added
+- Generator for dry-monad.
+- MONADS_PATH constant to Dry namespace.
+- dry-monads development dependency.
+
+### Changed
+- Version bumped.
+- Updated README.
+
 ## [0.3.0] - 2020-02-16
 ### Added
 - A generator for dry-struct.